--- src/main/java/com/example/BarrierKU/domain/Type/ImageType.java
@@ -1,6 +1,0 @@
-package com.example.BarrierKU.domain.Type;
-
-public enum ImageType {
-    DOOR,
-    ROOM
-}
