--- src/main/java/com/example/BarrierKU/domain/Type/ObstacleType.java
@@ -1,7 +1,0 @@
-package com.example.BarrierKU.domain.Type;
-
-public enum ObstacleType {
-    CURB,
-    RUNWAY,
-    STAIR
-}
