--- src/main/java/com/example/BarrierKU/domain/Type/Purpose.java
@@ -1,16 +1,0 @@
-package com.example.BarrierKU.domain.Type;
-
-public enum Purpose {
-    BANK,
-    POSTOFFICE,
-    CULTUER,
-    COPY,
-    STORE,
-    RESTAURANT,
-    CAFE,
-    CONVINIENCE,
-    KHUB,
-    KCUBE,
-    FOYER,
-    PARKINGLOT
-}
