--- src/main/java/com/example/BarrierKU/domain/Type/RoomType.java
@@ -1,6 +1,0 @@
-package com.example.BarrierKU.domain.Type;
-
-public enum RoomType {
-    FLAT,
-    STEPPED
-}
