+++ src/main/java/com/example/BarrierKU/domain/building/dto/BuildingResponse.java
@@ -1,0 +1,40 @@
+package com.example.BarrierKU.domain.building.dto;
+
+import com.example.BarrierKU.domain.indoor.Building;
+import com.example.BarrierKU.domain.indoor.Significant;
+import com.example.BarrierKU.domain.type.Purpose;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.List;
+import java.util.Set;
+
+@Getter
+@AllArgsConstructor
+public class BuildingResponse {
+    private Long id;
+
+    private int number;
+
+    private String name;
+
+    private String department;
+
+    private String image;
+
+    private List<DoorResponse> doors;
+
+    private Set<Purpose> FacilityPurposes;
+
+    private List<Significant> significants;
+
+    public static BuildingResponse from(Building building){
+        return new BuildingResponse(building.getId(), building.getNumber(),
+                building.getName(), building.getDepartment(),
+                building.getImage(),
+                building.getDoors().stream()
+                .map(DoorResponse::from)
+                .toList(),
+                building.getFacilityPurposes(), building.getSignificants());
+    }
+}
