+++ src/main/java/com/example/BarrierKU/domain/building/dto/DoorResponse.java
@@ -1,0 +1,25 @@
+package com.example.BarrierKU.domain.building.dto;
+
+import com.example.BarrierKU.domain.image.DoorImage;
+import com.example.BarrierKU.domain.indoor.Door;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+@AllArgsConstructor
+public class DoorResponse {
+    private Long id;
+
+    private boolean wheelchair;
+
+    private String significant;
+
+    private List<DoorImage> images;
+
+    public static DoorResponse from(Door door) {
+        return new DoorResponse(door.getId(), door.isWheelchair(), door.getSignificant(), door.getImages());
+    }
+
+}
