+++ src/main/java/com/example/BarrierKU/domain/building/dto/SpaceResponse.java
@@ -1,0 +1,48 @@
+package com.example.BarrierKU.domain.building.dto;
+
+import com.example.BarrierKU.domain.indoor.Room;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+@AllArgsConstructor
+public class SpaceResponse {
+    private String roomNumber;
+    private String roomName;
+    private boolean lecture;
+    private long capacity;
+    private float area;
+    private String roomComment;
+    private float floorSpace;
+    private String roomType;
+    private String department;
+    private String departmentNumber;
+    private RoomInfoResponse roomInfo;
+    private List<RoomImageResponse> images;
+
+    public static SpaceResponse of(Room room, int type) {
+        RoomInfoResponse info = (type == 1 && room.getRoomInfo() != null)
+                ? RoomInfoResponse.from(room.getRoomInfo()) : null;
+
+        List<RoomImageResponse> images = room.getRoomImages().stream().map(RoomImageResponse::from).toList();
+
+        return new SpaceResponse (
+                room.getRoomNumber(),
+                room.getRoomName(),
+                room.isLecture(),
+                room.getCapacity(),
+                room.getArea(),
+                room.getRoomComment(),
+                room.getFloorSpace(),
+                room.getRoomType().name(),
+                room.getDepartment(),
+                room.getDepartmentNumber(),
+                info,
+                images
+        );
+
+    }
+}
+
