+++ src/main/java/com/example/BarrierKU/domain/building/service/BuildingService.java
@@ -1,0 +1,22 @@
+package com.example.BarrierKU.domain.building.service;
+
+import com.example.BarrierKU.common.exception.BarrierKuException;
+import com.example.BarrierKU.domain.indoor.Building;
+import com.example.BarrierKU.domain.building.dto.BuildingResponse;
+import com.example.BarrierKU.domain.building.repository.BuildingRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import static com.example.BarrierKU.common.response.ResponseCode.BUILDING_NOT_FOUND;
+
+@Service
+@RequiredArgsConstructor
+public class BuildingService {
+    private final BuildingRepository buildingRepository;
+
+    public BuildingResponse findBuildingById(Long id) {
+        Building building = buildingRepository.findById(id)
+                .orElseThrow(() -> new BarrierKuException(BUILDING_NOT_FOUND));
+        return BuildingResponse.from(building);
+    }
+}
