+++ src/main/java/com/example/BarrierKU/domain/image/OutsideSignificantImage.java
@@ -1,0 +1,25 @@
+package com.example.BarrierKU.domain.image;
+
+import com.example.BarrierKU.domain.outdoor.OutsideSignificant;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+
+
+@Entity
+@Getter
+public class OutsideSignificantImage {
+
+    @Id @GeneratedValue
+    @Column(name = "outside_siginificant_image_id")
+    private Long id;
+
+    @Column(nullable = false)
+    @NotNull
+    private String url;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "outside_significant_id", nullable = false)
+    private OutsideSignificant outsideSignificant;
+
+}
