+++ src/main/java/com/example/BarrierKU/domain/image/SignificantImage.java
@@ -1,0 +1,24 @@
+package com.example.BarrierKU.domain.image;
+
+import com.example.BarrierKU.domain.indoor.Significant;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+
+
+@Entity
+@Getter
+public class SignificantImage {
+
+    @Id @GeneratedValue
+    @Column(name = "significant_image_id")
+    private Long id;
+
+    @Column(nullable = false)
+    @NotNull
+    private String url;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "significant_id")
+    private Significant significant;
+}
