+++ src/main/java/com/example/BarrierKU/domain/indoor/Significant.java
@@ -1,0 +1,19 @@
+package com.example.BarrierKU.domain.indoor;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+
+@Entity
+@Getter
+public class Significant {
+
+    @Id @GeneratedValue
+    @Column(name = "significant_id")
+    private Long id;
+
+    private String description;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "building_id", nullable = false)
+    private Building building;
+}
