+++ src/main/java/com/example/BarrierKU/domain/outdoor/Obstacle.java
@@ -1,0 +1,23 @@
+package com.example.BarrierKU.domain.outdoor;
+
+import com.example.BarrierKU.domain.type.ObstacleType;
+import jakarta.persistence.*;
+import lombok.Getter;
+import org.locationtech.jts.geom.Point;
+
+@Entity
+@Getter
+public class Obstacle {
+
+    @Id @GeneratedValue
+    @Column(name = "obstacle_id")
+    private Long id;
+
+    @Column(nullable = false)
+    private Point spot;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "type", nullable = false)
+    private ObstacleType obstacleType;
+}
+
