+++ src/main/java/com/example/BarrierKU/domain/type/ImageType.java
@@ -1,0 +1,6 @@
+package com.example.BarrierKU.domain.type;
+
+public enum ImageType {
+    DOOR,
+    ROOM
+}
