+++ src/main/java/com/example/BarrierKU/domain/type/ObstacleType.java
@@ -1,0 +1,7 @@
+package com.example.BarrierKU.domain.type;
+
+public enum ObstacleType {
+    CURB,
+    RUNWAY,
+    STAIR
+}
