+++ src/main/java/com/example/BarrierKU/domain/type/Purpose.java
@@ -1,0 +1,16 @@
+package com.example.BarrierKU.domain.type;
+
+public enum Purpose {
+    BANK,
+    POSTOFFICE,
+    CULTUER,
+    COPY,
+    STORE,
+    RESTAURANT,
+    CAFE,
+    CONVINIENCE,
+    KHUB,
+    KCUBE,
+    FOYER,
+    PARKINGLOT
+}
