+++ src/main/java/com/example/BarrierKU/domain/type/RoomType.java
@@ -1,0 +1,6 @@
+package com.example.BarrierKU.domain.type;
+
+public enum RoomType {
+    FLAT,
+    STEPPED
+}
