--- src/main/java/com/example/BarrierKU/repository/BuildingRepository.java
@@ -1,8 +1,0 @@
-package com.example.BarrierKU.repository;
-
-import com.example.BarrierKU.domain.Indoor.Building;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BuildingRepository extends JpaRepository<Building, Long> {
-
-}
