--- src/main/java/com/example/BarrierKU/service/SpaceService.java
@@ -1,25 +1,0 @@
-package com.example.BarrierKU.service;
-
-import com.example.BarrierKU.common.exception.BarrierKuException;
-import com.example.BarrierKU.domain.Indoor.Room;
-import com.example.BarrierKU.dto.SpaceResponse;
-import com.example.BarrierKU.repository.RoomRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import static com.example.BarrierKU.common.response.ResponseCode.SPACE_NOT_FOUND;
-
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class SpaceService {
-
-    private final RoomRepository roomRepository;
-
-    public SpaceResponse getSpaceInfo(Long buildingId, Long spaceId, int type) {
-        Room room = roomRepository.findByIdAndBuildingId(buildingId, spaceId)
-                .orElseThrow(() -> new BarrierKuException(SPACE_NOT_FOUND));
-        return SpaceResponse.of(room, type);
-    }
-}
